+++ CHANGELOG.md
@@ -28,6 +28,8 @@
   `moves` will only be set when a composed interaction has failed.
 - The `inputs` method now returns an `ActiveInteraction::Input` instead of a
   hash. The `ActiveInteraction::Input` still responds to all hash methods.
+- The `interface` filter will now look for an ancestor of the value passed
+  based on the name of the interface or the value passed in the `from` option.
 
 ## Added